requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
Django==3.0.3
entrypoints==0.3
flake8==3.7.9
future==0.18.2
isort==4.3.21
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-telegram-bot==12.4.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3
